메인화면 애니메이션 + 서버 타이머 수정 + 안정화 작업

+++ Server/Server/C_Timer.cpp
@@ -27,7 +27,7 @@ double C_Timer::ElapsedMilliseconds()
 		nowEndTime = endTime;
 	}
 
-	return duration_cast<milliseconds>(nowEndTime - startTime).count();
+	return (double)duration_cast<milliseconds>(nowEndTime - startTime).count();
 }
 
 double C_Timer::ElapsedSeconds()
